+++ lib/seasonal_decor.dart
@@ -11,7 +11,7 @@ export 'src/presets/eid.dart' show EidVariant;
 export 'src/presets/halloween.dart' show HalloweenVariant;
 export 'src/presets/new_year.dart' show NewYearVariant;
 export 'src/presets/ramadan.dart' show RamadanVariant;
-export 'src/presets/football.dart' show SportEventVariant;
+export 'src/presets/football.dart' show FootballVariant;
 export 'src/presets/seasonal_preset.dart';
 export 'src/presets/valentine.dart' show ValentineVariant;
 export 'src/widgets/seasonal_decor.dart';